refactor(components/modal-pages): to not use hooks yet
