+++ packages/application-components/src/components/modal-pages/internals/modal-page.js
@@ -1,17 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Modal from 'react-modal';
-import { css, ClassNames } from '@emotion/core';
+import { ClassNames } from '@emotion/core';
 import { PORTALS_CONTAINER_ID } from '@commercetools-frontend/constants';
-import { injectIntl } from 'react-intl';
-import {
-  Text,
-  CloseIcon,
-  FlatButton,
-  AngleLeftIcon,
-  SecondaryIconButton,
-  customProperties,
-} from '@commercetools-frontend/ui-kit';
 import {
   getOverlayStyles,
   getContainerStyles,
@@ -20,7 +11,7 @@ import {
   getBeforeCloseOverlayAnimation,
   getBeforeCloseContainerAnimation,
 } from './modal-page.styles';
-import messages from './messages';
+import ModalPageTopBar from './modal-page-top-bar';
 
 const TRANSITION_DURATION = 200;
 
@@ -42,214 +33,117 @@ const getDefaultParentSelector = () =>
     ? document.body
     : document.querySelector(`#${PORTALS_CONTAINER_ID}`);
 
-// Component to have a larger the clickable surface
-const LargeCloseIcon = props => (
-  <span
-    css={css`
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      &::after {
-        content: '';
-        position: absolute;
-        height: 35px;
-        width: 48px;
-        top: 0;
-        right: 0;
-      }
-    `}
-  >
-    <CloseIcon size="medium" {...props} />
-  </span>
-);
-LargeCloseIcon.displayName = 'LargeCloseIcon';
-
-const ModalPageTopBar = injectIntl(props => (
-  <div
-    css={css`
-      position: relative;
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      padding: ${customProperties.spacingS} ${customProperties.spacingM};
-      background-color: ${props.color === 'neutral'
-        ? customProperties.colorNeutral95
-        : customProperties.colorSurface};
-      border-bottom: 1px solid
-        ${props.color === 'neutral'
-          ? customProperties.colorSurface
-          : customProperties.colorNeutral};
-      & * + * {
-        margin-left: ${customProperties.spacingS};
-      }
-
-      /* FIXME: these "dirty" styles should be removed when the new Breadcrumbs component is implemented */
-      p {
-        font-size: 12px !important;
-      }
-
-      /*  specific selector for the svg of the FlatButton */
-      button > span > svg {
-        height: 12px !important;
-        width: 12px !important;
-      }
-    `}
-  >
-    <div
-      css={css`
-        display: flex;
-        overflow: hidden;
-      `}
-    >
-      <FlatButton
-        tone="primary"
-        label={
-          typeof props.previousPathLabel === 'string'
-            ? props.previousPathLabel
-            : props.intl.formatMessage(props.previousPathLabel)
-        }
-        icon={<AngleLeftIcon size="medium" theme="green" />}
-        onClick={props.onClose}
-      />
-      {props.currentPathLabel && (
-        <React.Fragment>
-          <Text.Detail isInline>/</Text.Detail>
-          <Text.Detail title={props.currentPathLabel} isInline truncate>
-            {props.currentPathLabel}
-          </Text.Detail>
-        </React.Fragment>
-      )}
-    </div>
-    {props.onClose && (
-      <SecondaryIconButton
-        label={props.intl.formatMessage(messages.close)}
-        onClick={props.onClose}
-        icon={<LargeCloseIcon />}
-      />
-    )}
-  </div>
-));
-ModalPageTopBar.displayName = 'ModalPageTopBar';
-ModalPageTopBar.propTypes = {
-  color: PropTypes.oneOf(['surface', 'neutral']),
-  onClose: PropTypes.func.isRequired,
-  currentPathLabel: PropTypes.string,
-  previousPathLabel: PropTypes.oneOfType([
-    PropTypes.string,
-    // default intl message
-    PropTypes.shape({
-      id: PropTypes.string.isRequired,
-      defaultMessage: PropTypes.string.isRequired,
+class ModalPage extends React.Component {
+  static displayName = 'ModalPage';
+  static propTypes = {
+    level: PropTypes.number,
+    title: PropTypes.string.isRequired,
+    zIndex: PropTypes.number,
+    isOpen: PropTypes.bool.isRequired,
+    onClose: PropTypes.func,
+    children: PropTypes.node.isRequired,
+    baseZIndex: PropTypes.number,
+    getParentSelector: PropTypes.func,
+    shouldDelayOnClose: PropTypes.bool,
+    // TopBar props:
+    topBarColor: PropTypes.oneOf(['surface', 'neutral']),
+    currentPathLabel: PropTypes.string,
+    previousPathLabel: PropTypes.oneOfType([
+      PropTypes.string,
+      // default intl message
+      PropTypes.shape({
+        id: PropTypes.string.isRequired,
+        defaultMessage: PropTypes.string.isRequired,
+      }),
+    ]).isRequired,
+    // injected
+    intl: PropTypes.shape({
+      formatMessage: PropTypes.func.isRequired,
     }),
-  ]).isRequired,
-  // injected
-  intl: PropTypes.shape({
-    formatMessage: PropTypes.func.isRequired,
-  }),
-};
-ModalPageTopBar.defaultProps = {
-  color: 'surface',
-  previousPathLabel: messages.back,
-};
-
-const ModalPage = props => {
-  const [isOpen, setIsOpen] = React.useState(props.isOpen);
-
-  const delayedOnClose = () => {
-    setIsOpen(false);
-    setTimeout(() => {
-      props.onClose();
-    }, TRANSITION_DURATION);
   };
-
-  if (props.shouldDelayOnClose) {
-    React.useEffect(() => {
-      setIsOpen(props.isOpen);
-    }, [props.isOpen]);
+  static defaultProps = {
+    level: 1,
+    baseZIndex: 1000,
+    getParentSelector: getDefaultParentSelector,
+    shouldDelayOnClose: true,
+  };
+  static getDerivedStateFromProps(props) {
+    return { isOpen: props.isOpen };
   }
 
-  React.useEffect(() => {
-    return () => {
-      clearTimeout(delayedOnClose);
-    };
-  }, []);
+  componentDidUpdate(prevProps) {
+    if (
+      this.props.shouldDelayOnClose &&
+      prevProps.isOpen !== this.props.isOpen
+    ) {
+      this.setState({ isOpen: this.props.isOpen });
+    }
+  }
 
-  return (
-    <ClassNames>
-      {({ css: makeClassName }) => (
-        <Modal
-          isOpen={props.shouldDelayOnClose ? isOpen : props.isOpen}
-          onRequestClose={
-            props.shouldDelayOnClose ? delayedOnClose : props.onClose
-          }
-          shouldCloseOnOverlayClick={Boolean(props.onClose)}
-          shouldCloseOnEsc={Boolean(props.onClose)}
-          overlayClassName={{
-            base: makeClassName(getOverlayStyles(props, TRANSITION_DURATION)),
-            afterOpen: makeClassName(getAfterOpenOverlayAnimation()),
-            beforeClose: makeClassName(getBeforeCloseOverlayAnimation()),
-          }}
-          className={{
-            base: makeClassName(getContainerStyles(props, TRANSITION_DURATION)),
-            afterOpen: makeClassName(getAfterOpenContainerAnimation()),
-            beforeClose: makeClassName(getBeforeCloseContainerAnimation()),
-          }}
-          contentLabel={props.title}
-          parentSelector={props.getParentSelector}
-          ariaHideApp={false}
-          // Adjust this value if the (beforeClose) animation duration is changed
-          closeTimeoutMS={TRANSITION_DURATION}
-          style={{
-            // stylelint-disable-next-line selector-type-no-unknown
-            overlay: {
-              zIndex: props.zIndex,
-            },
-          }}
-        >
-          <ModalPageTopBar
-            color={props.topBarColor}
-            onClose={props.shouldDelayOnClose ? delayedOnClose : props.onClose}
-            currentPathLabel={props.currentPathLabel}
-            previousPathLabel={props.previousPathLabel}
-          />
-          {props.children}
-        </Modal>
-      )}
-    </ClassNames>
-  );
-};
-ModalPage.displayName = 'ModalPage';
-ModalPage.propTypes = {
-  level: PropTypes.number,
-  title: PropTypes.string.isRequired,
-  zIndex: PropTypes.number,
-  isOpen: PropTypes.bool.isRequired,
-  onClose: PropTypes.func,
-  children: PropTypes.node.isRequired,
-  baseZIndex: PropTypes.number,
-  getParentSelector: PropTypes.func,
-  shouldDelayOnClose: PropTypes.bool,
-  // TopBar props:
-  topBarColor: PropTypes.oneOf(['surface', 'neutral']),
-  currentPathLabel: PropTypes.string,
-  previousPathLabel: PropTypes.oneOfType([
-    PropTypes.string,
-    // default intl message
-    PropTypes.shape({
-      id: PropTypes.string.isRequired,
-      defaultMessage: PropTypes.string.isRequired,
-    }),
-  ]).isRequired,
-  // injected
-  intl: PropTypes.shape({
-    formatMessage: PropTypes.func.isRequired,
-  }),
-};
-ModalPage.defaultProps = {
-  level: 1,
-  baseZIndex: 1000,
-  getParentSelector: getDefaultParentSelector,
-  shouldDelayOnClose: true,
-};
+  componentWillUnmount() {
+    if (this.closingTimer) {
+      clearTimeout(this.closingTimer);
+    }
+  }
+
+  handleClose = () => {
+    if (this.props.shouldDelayOnClose) {
+      this.setState({ isOpen: false });
+      this.closingTimer = setTimeout(() => {
+        this.props.onClose();
+      }, TRANSITION_DURATION);
+      return;
+    }
+    this.props.onClose();
+  };
+
+  render() {
+    return (
+      <ClassNames>
+        {({ css: makeClassName }) => (
+          <Modal
+            isOpen={this.state.isOpen}
+            onRequestClose={this.handleClose}
+            shouldCloseOnOverlayClick={Boolean(this.props.onClose)}
+            shouldCloseOnEsc={Boolean(this.props.onClose)}
+            overlayClassName={{
+              base: makeClassName(
+                getOverlayStyles(this.props, TRANSITION_DURATION)
+              ),
+              afterOpen: makeClassName(getAfterOpenOverlayAnimation()),
+              beforeClose: makeClassName(getBeforeCloseOverlayAnimation()),
+            }}
+            className={{
+              base: makeClassName(
+                getContainerStyles(this.props, TRANSITION_DURATION)
+              ),
+              afterOpen: makeClassName(getAfterOpenContainerAnimation()),
+              beforeClose: makeClassName(getBeforeCloseContainerAnimation()),
+            }}
+            contentLabel={this.props.title}
+            parentSelector={this.props.getParentSelector}
+            ariaHideApp={false}
+            // Adjust this value if the (beforeClose) animation duration is changed
+            closeTimeoutMS={TRANSITION_DURATION}
+            style={{
+              // stylelint-disable-next-line selector-type-no-unknown
+              overlay: {
+                zIndex: this.props.zIndex,
+              },
+            }}
+          >
+            <ModalPageTopBar
+              color={this.props.topBarColor}
+              onClose={this.handleClose}
+              currentPathLabel={this.props.currentPathLabel}
+              previousPathLabel={this.props.previousPathLabel}
+            />
+            {this.props.children}
+          </Modal>
+        )}
+      </ClassNames>
+    );
+  }
+}
 
 export default ModalPage;
